requirements.txt
boto3>=1.26.0
PyYAML>=6.0
openpyxl>=3.0.0
python-docx>=0.8.11
colorama>=0.4.4
semver>=3.0.0
psutil>=5.9.0
keyring>=24.0.0
cryptography>=41.0.0
requests>=2.31.0